refactor: move alternatePath function from the tmux package to the dir package
