+++ dir/paths_test.go
@@ -1,4 +1,4 @@
-package tmux
+package dir
 
 import (
 	"os"
@@ -9,16 +9,16 @@ import (
 
 func TestAlternatePath(t *testing.T) {
 	t.Run("absolute path", func(t *testing.T) {
-		require.Equal(t, "", alternatePath("/foo/bar"))
+		require.Equal(t, "", AlternatePath("/foo/bar"))
 	})
 	t.Run("home directory", func(t *testing.T) {
 		homeDir, err := os.UserHomeDir()
 		require.NoError(t, err)
-		require.Equal(t, homeDir+"/foo/bar", alternatePath("~/foo/bar"))
+		require.Equal(t, homeDir+"/foo/bar", AlternatePath("~/foo/bar"))
 	})
 	t.Run("relative path", func(t *testing.T) {
 		wd, err := os.Getwd()
 		require.NoError(t, err)
-		require.Equal(t, wd+"/foo/bar", alternatePath("./foo/bar"))
+		require.Equal(t, wd+"/foo/bar", AlternatePath("./foo/bar"))
 	})
 }
